Responding to CR feedback to DRY up ViewModel and SelectionViewModel
