+++ Examples/iTunesSearch/Shared/TelevisionEpisodeViewModel.swift
@@ -32,7 +32,7 @@ public struct TelevisionEpisodeViewModel: ViewModel {
 
     public let context: Context
 
-    public func action(_ event: ViewModelUserEvent) -> Action? {
+    public func actionForUserEvent(_ event: ViewModelUserEvent) -> Action? {
         if case .select = event {
             return ViewMediaAction(url: episode.preview)
         }
